+++ src/GIntervalsBigSet2D.h
@@ -1,6 +1,7 @@
 #ifndef _GINTERVALSBIGSET2D_H_INCLUDED_
 #define _GINTERVALSBIGSET2D_H_INCLUDED_
 
+#include <cstdint>
 #include "GIntervals2D.h"
 #include "GIntervalsBigSet.h"
 #include "GIntervalsFetcher2D.h"
@@ -56,14 +57,14 @@ public:
 	virtual bool next();
 	virtual bool next_in_chrom();
 
-	virtual bool isend() const { return m_iter_index == -1 || m_iter_index >= m_size; }
+	virtual bool isend() const { return (int)m_iter_index == -1 || m_iter_index >= m_size; }
 	virtual bool isend_chrom() const { return m_iinterval >= m_intervals.end() || m_cur_chromid != m_iter_chromid; }
 
 	virtual GIntervals2D::const_iterator get_chrom_begin() const { return m_intervals.begin(); }
 	virtual GIntervals2D::const_iterator get_chrom_end() const { return m_intervals.end(); }
 
-	virtual size_t iter_index() const { return m_iter_index; }
-	virtual size_t iter_chrom_index() const { return m_iter_chrom_index; }
+	virtual uint64_t iter_index() const { return m_iter_index; }
+	virtual uint64_t iter_chrom_index() const { return m_iter_chrom_index; }
 
 	virtual const GInterval2D &cur_interval() const { return *m_iinterval; }
 
@@ -76,8 +77,8 @@ private:
 	GIntervals2D::const_iterator m_iinterval;
 	int                          m_cur_chromid;
 	int                          m_iter_chromid;
-	size_t                       m_iter_index;
-	size_t                       m_iter_chrom_index;
+	uint64_t                       m_iter_index;
+	uint64_t                       m_iter_chrom_index;
 	Compare_t                    m_compare;
 	bool                         m_do_sort;
