+++ src/System.h
@@ -1,5 +1,6 @@
 #ifndef _SYSTEM_H_INCLUDED_
 #define _SYSTEM_H_INCLUDED_
+#include <cstdint>
 
 // This function returns the actual number of unique bytes (cannot shared with other processes) used by pid.
 // This is done by counting Private Dirty memory mappings.
@@ -10,6 +11,6 @@
 // Note: this function has high run-time cost since it needs to open and to parse system files.
 // 
 // !!!! COMPATIBILITY: only Linux, on other platforms returns 0. !!!!
-size_t get_unique_mem_usage(pid_t pid = 0);
+uint64_t get_unique_mem_usage(pid_t pid = 0);
 
 #endif
